Restructure to remove import loop.
