+++ juju/unit.py
@@ -2,7 +2,8 @@ import logging
 
 import pyrfc3339
 
-from . import annotation, model, tag
+from . import model, tag
+from .annotationhelper import _get_annotations, _set_annotations
 from .client import client
 
 log = logging.getLogger(__name__)
@@ -285,7 +286,7 @@ class Unit(model.ModelEntity):
 
         :return dict: The annotations for this unit
         """
-        return await annotation._get_annotations(self.tag, self.connection)
+        return await _get_annotations(self.tag, self.connection)
 
     async def set_annotations(self, annotations):
         """Set annotations on this unit.
@@ -294,5 +295,4 @@ class Unit(model.ModelEntity):
             pairs.
 
         """
-        return await annotation._set_annotations(
-            self.tag, annotations, self.connection)
+        return await _set_annotations(self.tag, annotations, self.connection)
